refactor: move new benchmark under stateful folder
